Add workaround for command parser bug
There is currently a bug in the command parser where a command ending in
the magic number (0xAA) will cause the device to expect a new command.
We add a 0x00 to the stop_outputs command bytestream to prevent this
from happening (Yuck!). Hopefully we won't run into this anywhere else.